use 64-bit index dual views for neighbor history and MLIAP
Two dual views hold more than 2^31 entries at atom counts people actually run.
fix_neigh_history's firstvalue is one entry per particle per neighbor per
history value, so with 40 neighbors and 7 values it passes the limit at 7.7
million particles per MPI rank, which is an ordinary granular simulation today;
firstflag and the partner arrays are the same shape without the history factor.
MLIAP's betas and descriptors are one entry per atom per descriptor over the
whole local list, with no chunking, so a SNAP descriptor at twojmax 8 reaches
the limit at 39 million atoms per rank.

Declaring these through KKBigDualView widens both the device and the host view,
and leaves everything else in the package indexed in 32 bits, which is faster on
GPUs.  The arrays that only overflow at seventy million atoms per rank and above
-- the molecular topology arrays, the minimizer vectors, the per-atom analysis
computes -- are left alone: reaching them needs a workload that does not exist.

memory_kokkos.h is a separate matter.  It sets up the CPU-side row pointers of
every array made by create_kokkos and grow_kokkos with

    array[i] = &data.view_host()(i,0);

which is itself a 32-bit view offset over the full extent, so it wraps at the
same sizes and every later access through those pointers is then wrong.  Plain
Memory::create accumulates the same offset in a bigint (memory.h:154), so this
is not the CPU behavior being matched but a difference from it.  The loop runs
once per grow, so widening the index there costs nothing.

Checked with an allocation of the shape that breaks, 8 million particles by 280
history entries: a 32-bit view of it is short by exactly 2^32 while the widened
one lands on the right element, on the host as well as the device, and what a
kernel writes through it is what the dual view holds.  examples/granular
in.pour.drum and in.pour.flatwall are identical to the CPU styles, the granular,
mliap and snap tests pass, and ctest is 896 of 898 as before.

+++ src/KOKKOS/memory_kokkos.h
@@ -116,7 +116,7 @@ TYPE create_kokkos(TYPE &data, typename TYPE::value_type **&array,
     if (n2 == 0)
       array[i] = nullptr;
     else
-      array[i] = &data.view_host()(i,0);
+      array[i] = &data.view_host()((bigint) i,0);
   }
   return data;
 }
@@ -164,7 +164,7 @@ TYPE create4d_offset_kokkos(TYPE &data, typename TYPE::value_type ****&array,
             if (n4 == 0)
               array[i][j][k] = nullptr;
             else
-              array[i][j][k] = &data.view_host()(i,j,k,0);
+              array[i][j][k] = &data.view_host()((bigint) i,j,k,0);
           }
         }
       }
@@ -211,7 +211,7 @@ TYPE create_kokkos(TYPE &data, typename TYPE::value_type **&array,
     if (data.view_host().extent(1) == 0)
       array[i] = nullptr;
     else
-      array[i] = &data.view_host()(i,0);
+      array[i] = &data.view_host()((bigint) i,0);
 
   return data;
 }
@@ -236,7 +236,7 @@ TYPE grow_kokkos(TYPE &data, typename TYPE::value_type **&array,
     if (n2 == 0)
       array[i] = nullptr;
     else
-      array[i] = &data.view_host()(i,0);
+      array[i] = &data.view_host()((bigint) i,0);
 
   return data;
 }
@@ -259,7 +259,7 @@ TYPE grow_kokkos(TYPE &data, typename TYPE::value_type **&array,
     if (data.view_host().extent(1) == 0)
       array[i] = nullptr;
     else
-      array[i] = &data.view_host()(i,0);
+      array[i] = &data.view_host()((bigint) i,0);
 
   return data;
 }
@@ -309,7 +309,7 @@ TYPE create_kokkos(TYPE &data, typename TYPE::value_type ***&array,
         if (n3 == 0)
            array[i][j] = nullptr;
          else
-           array[i][j] = &data.view_host()(i,j,0);
+           array[i][j] = &data.view_host()((bigint) i,j,0);
       }
     }
   }
@@ -343,7 +343,7 @@ template <typename TYPE, typename HTYPE>
         if (n3 == 0)
            array[i][j] = nullptr;
          else
-           array[i][j] = &data.view_host()(i,j,0);
+           array[i][j] = &data.view_host()((bigint) i,j,0);
       }
     }
   }
@@ -390,7 +390,7 @@ TYPE grow_kokkos(TYPE &data, typename TYPE::value_type ***&array,
         if (n3 == 0)
            array[i][j] = nullptr;
          else
-           array[i][j] = &data.view_host()(i,j,0);
+           array[i][j] = &data.view_host()((bigint) i,j,0);
       }
     }
   }
